ai-service/requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
python-dotenv==1.0.1

# LLM and AI dependencies
langchain==0.2.16
langchain-openai==0.1.25

# Database dependencies
pymongo==4.6.0
motor==3.3.2

# UI dependencies
gradio==4.44.0
requests==2.31.0

# Voice processing dependencies
openai>=1.3.0
python-multipart>=0.0.6
aiofiles>=23.2.0
